bons10 Getting a `Error publishing message`. Syntax issue. Will fix later
